JobsApplicationsResumes/requirements.txt
fastapi
uvicorn
pydantic
jinja2
pyyaml
glob2
spacy 